requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
transformers>=4.35.2
torch>=2.2.0
torchaudio>=2.2.0
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1
numpy>=1.26.2
python-dotenv==1.0.0
requests==2.31.0
pyquran==1.0.1
rapidfuzz==3.5.2
scipy==1.11.4
dtw-python==1.3.0
huggingface-hub==0.19.4
